SyncRedirector refactor
