+++ Addon for AnkiDesktop/SyncRedirector/Utils.py
@@ -1,0 +1,5 @@
+def url_ending_with_slash(url):
+    if url[-1] == "/":
+        return url
+    else:
+        return url + "/"
